feat : Redis 세팅 완료

+++ storage-service/src/main/java/shop/lotfresh/storageservice/domain/orderproduct/api/request/OrderRequest.java
@@ -20,7 +20,7 @@ public class OrderRequest {
 
     private long orderId;
 
-    private long stock;
+    private Integer stock;
 
     private int isDeleted;
