+++ storage-service/src/main/java/shop/lotfresh/storageservice/domain/orderproduct/api/request/ProductInfo.java
@@ -9,7 +9,7 @@ public class ProductInfo {
 
     private long productId;
 
-    private long stock;
+    private Integer stock;
 
     private int isDeleted;
 }
